+++ orator/commands/migrations/make_command.py
@@ -35,9 +35,13 @@ class MigrateMakeCommand(BaseCommand):
         if path is None:
             path = self._get_migration_path()
 
-        self._write_migration(creator, name, table, create, path)
+        migration_name = self._write_migration(
+            creator, name, table, create, path
+        )
 
-        self.info('<info>Migration created successfully</info>')
+        self.line(
+            '<info>Created migration:</info> {}'.format(migration_name)
+        )
 
     def _write_migration(self, creator, name, table, create, path):
         """