+++ orator/commands/migrations/refresh_command.py
@@ -12,25 +12,24 @@ class RefreshCommand(BaseCommand):
         {--p|path= : The path of migrations files to be executed.}
         {--s|seed : Indicates if the seed task should be re-run.}
         {--seed-path= : The path of seeds files to be executed.
-                        Defaults to <comment>./seeders</comment>.}
+                        Defaults to <comment>./seeds</comment>.}
         {--seeder=database_seeder : The name of the root seeder.}
+        {--f|force : Force the operation to run.}
     """
 
     def handle(self):
         """
         Executes the command.
         """
-        confirm = self.confirm(
-            '<question>Are you sure you want to refresh the database?</question> ',
-            False
-        )
-        if not confirm:
+        if not self.confirm_to_proceed(
+            '<question>Are you sure you want to refresh the database?:</question> '
+        ):
             return
 
         database = self.option('database')
 
         options = [
-            ('-n', True)
+            ('--force', True)
         ]
 
         if self.option('path'):
@@ -50,12 +49,12 @@ class RefreshCommand(BaseCommand):
             self._run_seeder(database)
 
     def _needs_seeding(self):
-        return self.option('seed') or self.option('seeder')
+        return self.option('seed')
 
     def _run_seeder(self, database):
         options = [
             ('--seeder', self.option('seeder')),
-            ('-n', True)
+            ('--force', True)
         ]
 
         if database: