+++ orator/query/processors/postgres_processor.py
@@ -44,6 +44,6 @@ class PostgresQueryProcessor(QueryProcessor):
         :type results: dict
 
         :return: The processed results
-        :return: dict
+        :return: list
         """
-        return map(lambda x: x['column_name'], results)
+        return list(map(lambda x: x['column_name'], results))