+++ tests/commands/migrations/test_reset_command.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import os
-from io import BytesIO
 from flexmock import flexmock
 from orator.migrations import Migrator
 from orator.commands.migrations import ResetCommand
 from orator import DatabaseManager
-from orator.connections import Connection
 from .. import OratorCommandTestCase
 
 
@@ -25,8 +23,9 @@ class ResetCommandTestCase(OratorCommandTestCase):
 
         command = flexmock(ResetCommand())
         command.should_receive('_get_config').and_return({})
+        command.should_receive('confirm').and_return(True)
 
-        self.run_command(command, input_stream=self.get_input_stream('y\n'))
+        self.run_command(command)
 
     def test_migration_can_be_pretended(self):
         resolver = flexmock(DatabaseManager)
@@ -41,8 +40,9 @@ class ResetCommandTestCase(OratorCommandTestCase):
 
         command = flexmock(ResetCommand())
         command.should_receive('_get_config').and_return({})
+        command.should_receive('confirm').and_return(True)
 
-        self.run_command(command, [('--pretend', True)], input_stream=self.get_input_stream('y\n'))
+        self.run_command(command, [('--pretend', True)])
 
     def test_migration_database_can_be_set(self):
         resolver = flexmock(DatabaseManager)
@@ -57,12 +57,22 @@ class ResetCommandTestCase(OratorCommandTestCase):
 
         command = flexmock(ResetCommand())
         command.should_receive('_get_config').and_return({})
+        command.should_receive('confirm').and_return(True)
 
-        self.run_command(command, [('--database', 'foo')], input_stream=self.get_input_stream('y\n'))
+        self.run_command(command, [('--database', 'foo')])
 
-    def get_input_stream(self, input_):
-        stream = BytesIO()
-        stream.write(input_.encode())
-        stream.seek(0)
+    def test_migration_can_be_forced(self):
+        resolver = flexmock(DatabaseManager)
+        resolver.should_receive('connection').and_return(None)
 
-        return stream
+        migrator_mock = flexmock(Migrator)
+        migrator_mock.should_receive('set_connection').once().with_args(None)
+        migrator_mock.should_receive('reset').once()\
+            .with_args(os.path.join(os.getcwd(), 'migrations'), False)\
+            .and_return(2)
+        migrator_mock.should_receive('get_notes').and_return([])
+
+        command = flexmock(ResetCommand())
+        command.should_receive('_get_config').and_return({})
+
+        self.run_command(command, [('--force', True)])