+++ tests/integrations/__init__.py
@@ -1,13 +1,44 @@
 # -*- coding: utf-8 -*-
 
+import os
+import json
+import logging
 import pendulum
 import simplejson as json
-from datetime import datetime, timedelta
+
+from datetime import datetime, timedelta, date
+from backpack import collect
 from orator import Model, Collection, DatabaseManager
 from orator.orm import morph_to, has_one, has_many, belongs_to_many, morph_many, belongs_to, scope, accessor
 from orator.orm.relations import BelongsToMany
 from orator.exceptions.orm import ModelNotFound
-from orator.exceptions.query import QueryException
+
+
+logger = logging.getLogger('orator.connection.queries')
+logger.setLevel(logging.DEBUG)
+
+
+class LoggedQueriesFormatter(logging.Formatter):
+
+    def __init__(self, fmt=None, datefmt=None, style='%'):
+        super(LoggedQueriesFormatter, self).__init__()
+
+        self.logged_queries = []
+
+    def format(self, record):
+        self.logged_queries.append(record.query)
+
+        return super(LoggedQueriesFormatter, self).format(record)
+
+    def reset(self):
+        self.logged_queries = []
+
+
+formatter = LoggedQueriesFormatter()
+handler = logging.StreamHandler(open(os.devnull, 'w'))
+
+handler.setFormatter(formatter)
+logger.addHandler(handler)
 
 
 class IntegrationTestCase(object):
@@ -30,7 +61,10 @@ class IntegrationTestCase(object):
             'database': ':memory:'
         }
 
-        return DatabaseManager(config)
+        db = DatabaseManager(config)
+        db.connection().enable_query_log()
+
+        return db
 
     @classmethod
     def tearDownClass(cls):
@@ -43,6 +77,8 @@ class IntegrationTestCase(object):
     def setUp(self):
         self.migrate()
         self.migrate('test')
+
+        formatter.reset()
 
     def tearDown(self):
         self.revert()
@@ -155,13 +191,32 @@ class IntegrationTestCase(object):
 
     def test_basic_has_many_eager_loading(self):
         user = OratorTestUser.create(id=1, email='john@doe.com')
-        user.posts().create(name='First Post')
-        user = OratorTestUser.with_('posts').where('email', 'john@doe.com').first()
+        post = user.posts().create(name='First Post')
+        comment = post.comments().create(body='Text')
+        comment2 = post.comments().create(body='Text 2')
+        comment.children().save(comment2)
+        user = OratorTestUser.with_('posts.comments.children.parent').where('email', 'john@doe.com').first()
 
         self.assertEqual('First Post', user.posts.first().name)
+        self.assertEqual('Text', user.posts.first().comments.first().body)
+        self.assertEqual('Text 2', user.posts.first().comments.first().children.first().body)
+        self.assertEqual('Text', user.posts.first().comments.first().children.first().parent.body)
+
+        queries = formatter.logged_queries
+        self.assertEqual(10, len(queries))
+
+        formatter.reset()
 
         post = OratorTestPost.with_('user').where('name', 'First Post').get()
         self.assertEqual('john@doe.com', post.first().user.email)
+
+        comment = OratorTestComment.with_('parent.post.user').where('body', 'Text 2').first()
+        self.assertEqual('Text', comment.parent.body)
+        self.assertEqual('First Post', comment.parent.post.name)
+        self.assertEqual('john@doe.com', comment.parent.post.user.email)
+
+        queries = formatter.logged_queries
+        self.assertEqual(6, len(queries))
 
     def test_basic_morph_many_relationship(self):
         user = OratorTestUser.create(email='john@doe.com')
@@ -194,11 +249,13 @@ class IntegrationTestCase(object):
 
     def test_multi_insert_with_different_values(self):
         date = pendulum.utcnow()._datetime
+        user1 = OratorTestUser.create(email='john@doe.com')
+        user2 = OratorTestUser.create(email='jane@doe.com')
         result = OratorTestPost.insert([
             {
-                'user_id': 1, 'name': 'Post', 'created_at': date, 'updated_at': date
+                'user_id': user1.id, 'name': 'Post', 'created_at': date, 'updated_at': date
             }, {
-                'user_id': 2, 'name': 'Post', 'created_at': date, 'updated_at': date
+                'user_id': user2.id, 'name': 'Post', 'created_at': date, 'updated_at': date
             }
         ])
 
@@ -207,11 +264,12 @@ class IntegrationTestCase(object):
 
     def test_multi_insert_with_same_values(self):
         date = pendulum.utcnow()._datetime
+        user1 = OratorTestUser.create(email='john@doe.com')
         result = OratorTestPost.insert([
             {
-                'user_id': 1, 'name': 'Post', 'created_at': date, 'updated_at': date
+                'user_id': user1.id, 'name': 'Post', 'created_at': date, 'updated_at': date
             }, {
-                'user_id': 1, 'name': 'Post', 'created_at': date, 'updated_at': date
+                'user_id': user1.id, 'name': 'Post', 'created_at': date, 'updated_at': date
             }
         ])
 
@@ -231,9 +289,13 @@ class IntegrationTestCase(object):
         self.assertIn('pivot', related_friend.to_dict())
         self.assertEqual(1, related_friend.pivot.user_id)
         self.assertEqual(3, related_friend.pivot.friend_id)
-        self.assertTrue(hasattr(related_friend.pivot, 'id'))
+        self.assertTrue(hasattr(related_friend.pivot, 'is_close'))
 
-        self.assertIsInstance(user.friends().with_pivot('id'), BelongsToMany)
+        self.assertIsInstance(user.friends().with_pivot('is_close'), BelongsToMany)
+
+        self.assertEqual(2, user.friends().get().count())
+        user.friends().sync([friend.id])
+        self.assertEqual(1, user.friends().get().count())
 
     def test_belongs_to_morph_many_eagerload(self):
         user = OratorTestUser.create(id=1, email='john@doe.com')
@@ -248,6 +310,27 @@ class IntegrationTestCase(object):
         self.assertEqual(user.id, posts[0].user().first().id)
         self.assertIsInstance(posts[0].photos, Collection)
         self.assertEqual(posts[0].photos().where('name', 'Hero 2').first().name, 'Hero 2')
+
+    def test_belongs_to_associate(self):
+        user = OratorTestUser.create(id=1, email='john@doe.com')
+        post = OratorTestPost(name='Test Post')
+
+        post.user().associate(user)
+        post.save()
+
+        self.assertEqual(1, post.user.id)
+
+    def test_belongs_to_associate_new_instances(self):
+        user = OratorTestUser.create(email='john@doe.com')
+        post = user.posts().create(name='First Post')
+        comment1 = OratorTestComment.create(body='test1', post_id=post.id)
+
+        self.assertEqual(comment1.parent, None)
+
+        comment2 = OratorTestComment.create(body='test2', post_id=post.id)
+        comment2.parent().associate(comment1)
+
+        self.assertEqual(comment2.parent.id, comment1.id)
 
     def test_has_many_eagerload(self):
         user = OratorTestUser.create(id=1, email='john@doe.com')
@@ -311,7 +394,7 @@ class IntegrationTestCase(object):
         self.assertEqual('bar', photo.metadata['foo'])
 
     def test_local_scopes(self):
-        yesterday = created_at=datetime.utcnow() - timedelta(days=1)
+        yesterday = datetime.utcnow() - timedelta(days=1)
         john = OratorTestUser.create(id=1, email='john@doe.com', created_at=yesterday, updated_at=yesterday)
         jane = OratorTestUser.create(id=2, email='jane@doe.com')
 
@@ -388,6 +471,14 @@ class IntegrationTestCase(object):
         self.assertEqual(1, users[0].id)
         self.assertEqual(1, users[0]['id'])
 
+    def test_query_builder_results_serialization(self):
+        OratorTestUser.create(id=1, email='john@doe.com')
+        users = self.connection().table('test_users').get()
+
+        serialized = json.loads(users.to_json())[0]
+        self.assertEqual(1, serialized['id'])
+        self.assertEqual('john@doe.com', serialized['email'])
+
     def test_connection_switching(self):
         OratorTestUser.create(id=1, email='john@doe.com')
 
@@ -399,6 +490,100 @@ class IntegrationTestCase(object):
         connection = user.get_connection()
         post = user.posts().create(name='Test')
         self.assertEqual(connection, post.get_connection())
+
+    def test_columns_listing(self):
+        column_names = (
+            collect(self.schema().get_column_listing(OratorTestUser().get_table()))
+            .sort()
+            .all()
+        )
+
+        self.assertEqual(['created_at', 'email', 'id', 'updated_at'], column_names)
+
+    def test_has_column(self):
+        self.assertTrue(
+            self.schema().has_column(OratorTestUser().get_table(), 'email')
+        )
+
+    def test_table_exists(self):
+        self.assertTrue(
+            self.schema().has_table(OratorTestUser().get_table())
+        )
+
+    def test_transaction(self):
+        count = self.connection().table('test_users').count()
+
+        with self.connection().transaction():
+            OratorTestUser.create(id=1, email='jane@doe.com')
+            self.connection().rollback()
+
+        self.assertEqual(count, self.connection().table('test_users').count())
+
+    def test_date(self):
+        user = OratorTestUser.create(id=1, email='john@doe.com')
+        photo1 = user.photos().create(name='Photo 1', taken_on=pendulum.date.today())
+        photo2 = user.photos().create(name='Photo 2')
+
+        self.assertIsInstance(OratorTestPhoto.find(photo1.id).taken_on, date)
+        self.assertIsNone(OratorTestPhoto.find(photo2.id).taken_on)
+
+    def test_chunk_update_builder(self):
+        for i in range(20):
+            self.connection().table('test_users').insert(id=i + 1, email='john{}@doe.com'.format(i))
+
+        count = 0
+        for users in self.connection().table('test_users').where('id', '<', 50).chunk(10):
+            for user in users:
+                count += 1
+
+                if count == 10:
+                    self.connection().table('test_users').where('id', user.id).update(id=60)
+
+        self.assertEqual(count, 20)
+
+    def test_chunk_update_model(self):
+        for i in range(20):
+            OratorTestUser.create(id=i + 1, email='john{}@doe.com'.format(i))
+
+        count = 0
+        for users in OratorTestUser.where('id', '<', 50).chunk(10):
+            for user in users:
+                count += 1
+
+                if count == 10:
+                    OratorTestUser.where('id', user.id).update(id=60)
+
+        self.assertEqual(count, 20)
+
+    def test_timestamp_with_timezone(self):
+        now = pendulum.utcnow()
+        user = OratorTestUser.create(email='john@doe.com', created_at=now)
+        fresh_user = OratorTestUser.find(user.id)
+
+        self.assertEqual(user.created_at, fresh_user.created_at)
+        self.assertEqual(now, fresh_user.created_at)
+
+    def test_touches(self):
+        user = OratorTestUser.create(email='john@doe.com')
+        post = user.posts().create(name='Post')
+        comment1 = post.comments().create(body='Comment 1')
+        comment2 = post.comments().create(body='Comment 2')
+        comment3 = post.comments().create(body='Comment 3')
+        comment4 = comment3.children().create(body='Comment 4', post_id=post.id)
+
+        comment1_updated_at = comment1.updated_at
+        comment2_updated_at = comment2.updated_at
+        comment3_updated_at = comment3.updated_at
+        comment4_updated_at = comment4.updated_at
+
+        comment4.body = 'Comment 4 updated'
+        comment4.save()
+
+        self.assertTrue(comment4.updated_at > comment4_updated_at)
+        self.assertEqual(comment4.updated_at, OratorTestComment.find(comment4.id).updated_at)
+        self.assertTrue(comment3_updated_at < OratorTestComment.find(comment3.id).updated_at)
+        self.assertEqual(comment1_updated_at, OratorTestComment.find(comment1.id).updated_at)
+        self.assertEqual(comment2_updated_at, OratorTestComment.find(comment2.id).updated_at)
 
     def grammar(self):
         return self.connection().get_default_query_grammar()
@@ -422,28 +607,45 @@ class IntegrationTestCase(object):
 
         with self.schema(connection).create('test_friends') as table:
             table.increments('id')
-            table.integer('user_id')
-            table.integer('friend_id')
+            table.integer('user_id').unsigned()
+            table.integer('friend_id').unsigned()
             table.boolean('is_close').default(False)
+
+            table.foreign('user_id').references('id').on('test_users').on_delete('cascade')
+            table.foreign('friend_id').references('id').on('test_users').on_delete('cascade')
 
         with self.schema(connection).create('test_posts') as table:
             table.increments('id')
-            table.integer('user_id')
+            table.integer('user_id').unsigned()
             table.string('name')
             table.timestamps(use_current=True)
+
+            table.foreign('user_id').references('id').on('test_users').on_delete('cascade')
+
+        with self.schema(connection).create('test_comments') as table:
+            table.increments('id')
+            table.integer('post_id').unsigned()
+            table.integer('parent_id').unsigned().nullable()
+            table.text('body')
+            table.timestamps(use_current=True)
+
+            table.foreign('post_id').references('id').on('test_posts').on_delete('cascade')
+            table.foreign('parent_id').references('id').on('test_comments').on_delete('cascade')
 
         with self.schema(connection).create('test_photos') as table:
             table.increments('id')
             table.morphs('imageable')
             table.string('name')
             table.json('metadata').nullable()
+            table.date('taken_on').nullable()
             table.timestamps(use_current=True)
 
     def revert(self, connection=None):
-        self.schema(connection).drop_if_exists('test_users')
-        self.schema(connection).drop_if_exists('test_friends')
-        self.schema(connection).drop_if_exists('test_posts')
         self.schema(connection).drop_if_exists('test_photos')
+        self.schema(connection).drop_if_exists('test_comments')
+        self.schema(connection).drop_if_exists('test_posts')
+        self.schema(connection).drop_if_exists('test_friends')
+        self.schema(connection).drop_if_exists('test_users')
 
     def get_marker(self):
         return '?'
@@ -454,7 +656,7 @@ class OratorTestUser(Model):
     __table__ = 'test_users'
     __guarded__ = []
 
-    @belongs_to_many('test_friends', 'user_id', 'friend_id', with_pivot=['id', 'is_close'])
+    @belongs_to_many('test_friends', 'user_id', 'friend_id', with_pivot=['is_close'])
     def friends(self):
         return OratorTestUser
 
@@ -472,7 +674,7 @@ class OratorTestUser(Model):
 
     @scope
     def older_than(self, query, **kwargs):
-        query.where('updated_at', '<', (pendulum.utcnow() - timedelta(**kwargs))._datetime)
+        query.where('updated_at', '<', pendulum.utcnow().subtract(**kwargs))
 
 
 class OratorTestPost(Model):
@@ -484,9 +686,33 @@ class OratorTestPost(Model):
     def user(self):
         return OratorTestUser
 
+    @has_many('post_id')
+    def comments(self):
+        return OratorTestComment
+
     @morph_many('imageable')
     def photos(self):
         return OratorTestPhoto.order_by('name')
+
+
+class OratorTestComment(Model):
+
+    __touches__ = ['parent']
+
+    __table__ = 'test_comments'
+    __guarded__ = []
+
+    @belongs_to('post_id')
+    def post(self):
+        return OratorTestPost
+
+    @belongs_to('parent_id')
+    def parent(self):
+        return OratorTestComment
+
+    @has_many('parent_id')
+    def children(self):
+        return OratorTestComment
 
 
 class OratorTestPhoto(Model):
@@ -498,10 +724,12 @@ class OratorTestPhoto(Model):
         'metadata': 'json'
     }
 
+    __dates__ = ['taken_on']
+
     @morph_to
     def imageable(self):
         return
 
     @accessor
     def created_at(self):
-        return pendulum.instance(self._attributes['created_at']).to('Europe/Paris')
+        return pendulum.instance(self._attributes['created_at']).in_tz('Europe/Paris')